Extract Getting Book from GoodReads in other class

+++ Library.Web.Shared/Author.cs
@@ -3,11 +3,10 @@
 	public class Author
 	{
 		public int Id { get; set; }
-		public int GoodReadsId { get; set; }
+		public long GoodReadsId { get; set; }
 		public string Name { get; set; }
 		public string ImageUrl { get; set; }
 		public string SmallImageUrl { get; set; }
 		public string Link { get; set; }
-
 	}
 }
